Triangular number matching - Add script and regexes
shortest - This was developed by teukon. The main algorithm inside was developed by me, but the factor capture method, and the method of doing the test for divisibility by N at the end at the same time as the check that the main algorithm returns a result of zero, was developed by teukon.

short and fast - Same factor-capturing technique as "shortest", but using a faster algorithm which is only 4 characters longer.

The requirement for both of these is portability (i.e., works with both ECMAScript and non-ECMAScript non-participating capture group behavior).