Fibonacci regex, fast: Move parity into first capture group; length+=6
This allows the multiplication optimization to be done more naturally. A way may be found later to make up for the added length.

This change has virtually no effect on speed.

Note, this change allows the main algorithm to successfully match 13, so this has been removed from the bottom, making it 2 characters shorter there.