Fibonacci regex: Add new method, which is hugely faster than any of the others
This is based on an algorithm idea formulated by teukon. I have put my own touches on it.

I wrote this entirely with speed optimization in mind. It is not at all optimized for length. The idea is that after committing the first working version, I can go through and find what length optimizations don't impact speed.