Abundant regex: Correct an incorrect assumption; length-=3
A needless extra comparison was done because I was incorrectly assuming that the next lower prime would be \3 if an exception were not made for the first iteration of the loop. But this is incorrect.

In an earlier version of the regex, it actually was necessary to seed the loop. Possibly this was before the initial commit of this regex.