prime^2 to prime index: Initial commit (already golfed)
This regex takes a squared prime as input, and matches its prime index, where {2,3,5,7,11,...} map to {1,2,3,4,5,...}.

I suspect that it is impossible to do the same thing for a non-squared prime, but am unsure how to prove this. Nor am I sure if it's possible to do for a prime, followed by delimiter, followed by scratch space, in less scratch space than O(p^2).