div sqrt2 regex: Add an interesting comment
The statement is true because of the use of ceiling-rounded sqrt(N), and was not true with floor+1 rounding. This coolness is the main reason I switched to ceiling rounding, despite the length+=8 cost it ended up having.

See also b005d2f3713900212c0b2451c4912e3c2c1b605b and the two subsequent commits.