Add regex matching factorial numbers; length = 739
The essence of how this algorithm works: It takes a guess at what the inverse factorial of the input is, and then compares the multiplicity of each prime factor of the input with what it would be if it were the factorial of the guessed number. The guess cycles through all possibilities and if the multiplicities all match, then it is factorial, and the inverse factorial is returned as a match.

This regex uses ECMAScript + molecular lookahead. It contains 3 molecular lookaheads. The first one would be challenging to remove, and would make the regex much longer. The next two should not be too hard to remove, and they made initial design of the regex easier (I could focus on getting it to work, rather than how to shuffle the space around to make up for the lack of molecular lookahead).