Add a new regex matching factorial numbers; length = 155
It currently does not match 1! or 2!, but matches 3! and above just fine. The method is hugely simpler than the previous one, and I wonder why I didn't think of it earlier.

It takes a guess M at what the inverse factorial of the input N is (guessing that M! == N), divides N by 2 to make room for the work it needs to do, and then seeds a loop in which it will repeatedly divide the input by a divisor that starts at 3 and increments each time. The divisor is kept track of by adding it to the running quotient; these two numbers can be unambiguously separated because, assuming M! == N, the running quotient will continue to be divisible by M until it equals M.

This regex uses ECMAScript + molecular lookahead, but it has only 1 molecular lookahead which should not be too difficult to remove.