factorial regex: Exponential slowdown; length-=1
Although the goal here is technically to golf the regex for length, with speed not being a consideration at all, this change is worse than any other I have ever golfed.

Theoretically I could implement an optimization for this in my engine, emulating the behavior of this construct without doing all the needless repetitive backtracking, but that would be a bit of a challenge, because the order in which ((x*)+) iterates through the possibilities is rather haphazard, and emulation of that order would need to be precise to guarantee the robustness of the optimization under all circumstances.

So I will abandon this, for the time being at least, as a dead end.