factorial regex: Use molecular lookaheads and return the full input as the match; length+=5
I was curious to see if bringing molecular lookaheads back into this algorithm, thus dropping the difference-of-3, would make it any shorter, and didn't know what to expect.

As it turned out, this resulted in a lengthening, even after I dropped the part of the specification that requires the regex to return the inverse factorial as a match. I think is really interesting; it means the difference-of-3 optimization is just that good.