Abundant regex, shortest: Reformat to fit in a cramped codegolf.stackexchange.com code box without needing (much) horizontal scrolling
The comments have also been changed from (?#...) to #... format.

This is done as an additional file because I don't want the primary one to have to be constrained to such a cramped width, and I dislike two-space indents. It's a very suboptimal solution to the problem, but it'll have to do for now.