Fibonacci number matching - Add a regex using a new method
This one builds up the Fibonacci numbers recursively from F_0=0, F_1=1 to F_n, the approximate square root of N, then uses identities to get F_{2n} and F_{2n-1}.

This bears some similarity to "regex for matching Fibonacci numbers - return index.txt" in that it builds up the Fibonacci numbers recursively, but that regex builds them up to the approximate fourth root of N, and thus needs to apply more identities to get N.

I'm not sure if this will be any use. It turned out to be longer than the 5a^2 ± 4 perfect square method, and it seems unlikely to be able to golfed down to a smaller size than it. And it is slower than the fourth root method, and unlike it, cannot return the Fibonacci index as a match. But something may be learned from golfing it down nevertheless.