div sqrt2 regex: Switch from ceiling to floor+1 rounding for calculating sqrt(N); length unchanged
It is due to the second change in 711a514cd6034ab099ce24870afbe82b09ea9c2a that this has no net effect on the total length of the regex. So I'm making this commit on a branch for illustrative purposes, but sticking with the ceiling rounding for the mainline branch.